[INTERPRETER] * NEW: Removes some warnings with uid_t on some systems.
[GB.EVAL]
* BUG: Removes the use of a non existing function.

[GB.DESKTOP]
* BUG: Try to enhance the detection of libXtst.


git-svn-id: svn://localhost/gambas/trunk@1220 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbx/gbx_c_file.c
@@ -267,7 +267,7 @@ static char *get_file_user(CFILE *_object)
     pwd = getpwuid(uid);
     if (!pwd)
     {
-      snprintf(_buffer, sizeof(_buffer), "%d", uid);
+      snprintf(_buffer, sizeof(_buffer), "%d", (int)uid);
       return _buffer;
     }
     else
@@ -294,7 +294,7 @@ static char *get_file_group(CFILE *_object)
     grp = getgrgid(gid);
     if (!grp)
     {
-      snprintf(_buffer, sizeof(_buffer), "%d", gid);
+      snprintf(_buffer, sizeof(_buffer), "%d", (int)gid);
       return _buffer;
     }
     else
